Add files via upload
This is the simple project in Python language for using the IR or the infrared sensor on the Raspberry pi. You can infer the pin number for vcc, ground and the output pin, set on the raspberry pi from the code. The variables have set are clearly understandable.
The projects teels the user using the infrared sensor whether there is any obstacle or not in its path. Modifying it will let the user to perform various function when there is an obstacle in its path or not.