Add files via upload
This Arduino program is used to blink led with different delay times, using a keypad. It has been made to make the users understand proper usage of keypads, and explore. Make sure that you experiment with various values to blink LED in different ways. And don't just restrict yourself to the LEDs.